// src/vizoApi/models/JoinType.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * An enumeration.
 */
export enum JoinType {
    INNER_JOIN = 'INNER JOIN',
    LEFT_JOIN = 'LEFT JOIN',
    RIGHT_JOIN = 'RIGHT JOIN',
    FULL_OUTER_JOIN = 'FULL OUTER JOIN',
}